Improve github username extraction
This commit creates a function for detecting whether a URL is a link
to a GitHub user, and returning the GitHub user's username. It takes
advantage of the built-in `url.parse` function to a greater degree than
the code it replaces, so as to reduce bugs and edge cases. The checks
are largely based on work in #49.

This commit also factors out the github username sanity checks in a
similar way.